+++ spec/models/partner.rb
@@ -1,0 +1,7 @@
+class Partner
+  include Mongoid::Document
+
+  field :name
+
+  has_many :artworks
+end