+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 Next Release
 ------------
 
+* [#18](https://github.com/aaw/mongoid_collection_snapshot/pull/18): Fix: classes that include `Mongoid::CollectionSnapshot` can be inherited - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 1.2.0
 -----
 
-* Compatibility with Mongoid 5 - [@dblock](https://github.com/dblock).
+* [#14](https://github.com/aaw/mongoid_collection_snapshot/pull/14): Compatibility with Mongoid 5 - [@dblock](https://github.com/dblock).
 
 1.1.0
 -----